live-code/requirements.txt
Flask==1.1.1
pymongo==3.10.1
requests==2.22.0
python-dotenv==0.10.5
marshmallow==3.3.0
